+++ InfiniteFriends/AssemblyInfo.cs
@@ -1,5 +1,5 @@
 using System.Reflection;
 
-[assembly: AssemblyVersion(InfiniteFriends.PluginInfo.PLUGIN_VERSION)]
-[assembly: AssemblyFileVersion(InfiniteFriends.PluginInfo.PLUGIN_VERSION)]
-[assembly: AssemblyInformationalVersion(InfiniteFriends.PluginInfo.PLUGIN_VERSION)]
+[assembly: AssemblyVersion(InfiniteFriends.Metadata.PLUGIN_VERSION)]
+[assembly: AssemblyFileVersion(InfiniteFriends.Metadata.PLUGIN_VERSION)]
+[assembly: AssemblyInformationalVersion(InfiniteFriends.Metadata.PLUGIN_VERSION)]