+++ InfiniteFriends/Metadata.cs
@@ -1,0 +1,24 @@
+namespace InfiniteFriends
+{
+    /// <summary>
+    /// Compile-time plugin metadata.
+    /// </summary>
+    public static class Metadata
+    {
+        /// <summary>
+        /// Plugin GUID, using reverse domain name notation.
+        /// </summary>
+        public const string PLUGIN_GUID       = "senyksia.spiderheck.infinitefriends";
+
+        /// <summary>
+        /// Human-readable name for logging, etc.
+        /// </summary>
+        public const string PLUGIN_NAME       = "InfiniteFriends";
+        public const string PLUGIN_NAME_SHORT = "InF";
+
+        /// <summary>
+        /// Current plugin version.
+        /// </summary>
+        public const string PLUGIN_VERSION    = "0.2.1";
+    }
+}